requirements.txt
selenium
openpyxl
keyboard
webdriver-manager
pyyaml
colorlog
pytest
pytest-mock
coverage
